Проектная-11

+++ mesto-react/src/utils/api.js
@@ -66,27 +66,43 @@ class Api {
     }).then(thenApi)
   }
 
-  plusLike(userId) {
-    return fetch(`${this._baseUrl}/cards/likes/${userId}`, {
-      method: 'PUT',
-      headers: this._headers,
-      body: JSON.stringify({
-        _id: userId,
-      })
-    }).then(thenApi)
-  }
+  // plusLike(userId) {
+  //   return fetch(`${this._baseUrl}/cards/likes/${userId}`, {
+  //     method: 'PUT',
+  //     headers: this._headers,
+  //   }).then(thenApi)
+  // }
 
-  deleteLike(userId) {
-    return fetch(`${this._baseUrl}/cards/likes/${userId}`, {
-      method: 'DELETE',
-      headers: this._headers,
-      body: JSON.stringify({
-        _id: userId,
-      })
-    }).then(thenApi)
-  }
+  // deleteLike(userId) {
+  //   return fetch(`${this._baseUrl}/cards/likes/${userId}`, {
+  //     method: 'DELETE',
+  //     headers: this._headers,
+  //     body: JSON.stringify({
+  //       _id: userId,
+  //     })
+  //   }).then(thenApi)
+  // }
 
+  changeLikeCardStatus(cardId, isLiked) {
+
+    if (isLiked) {
+      // this.plusLike(cardId);
+      return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+        method: 'PUT',
+        headers: this._headers
+      }).then(thenApi)
+    } else {
+      // this.deleteLike(cardId);
+      return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+        method: 'DELETE',
+        headers: this._headers
+      }).then(thenApi)
+    }
+
+
+  }
 }
+
 
 const api = new Api({
   baseUrl: 'https://mesto.nomoreparties.co/v1/cohort-17',
